backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# AI和ML
langchain==0.1.0
langchain-openai==0.0.2
openai==1.6.1
chromadb==0.4.18
sentence-transformers>=2.3.0
tiktoken==0.5.2

# 文档处理
pypdf==3.17.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
pandas==2.1.3

# 数据库连接器
pymysql==1.1.0
aiomysql==0.2.0
pymongo==4.6.0
motor==3.3.2

# 网络请求
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.1
cryptography==41.0.7

# 工具库
python-dotenv==1.0.0
pyyaml==6.0.1
orjson==3.9.10
aiofiles==23.2.1

# 日志和监控
loguru==0.7.2
sentry-sdk==1.38.0

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# 类型检查
mypy==1.7.1
types-redis==4.6.0.11

# 代码质量
black==23.11.0
flake8==6.1.0
isort==5.12.0

# 其他
celery==5.3.4
flower==2.0.1
tenacity==8.2.3

